presentation-builder-service/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# PPTX manipulation
python-pptx==0.6.23

# Storage
minio==7.2.3

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Service Discovery
python-consul==1.1.0

# JWT
pyjwt[crypto]==2.8.0
cryptography==41.0.7

# Config
pyyaml==6.0.1
python-dotenv==1.0.0

# Utils
python-multipart==0.0.6
